green - SetExpectedLines_long

+++ RayTracingApp/Engine/Progress.cs
@@ -9,7 +9,7 @@ namespace Engine
     public class Progress
     {
         public long LinesCount { get; set; }
-        public uint ExpectedLines { get; set; }
+        public long ExpectedLines { get; set; }
 
         public void Count()
         {
